+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## `1.0.48-SNAPSHOT`
+## `1.0.58-SNAPSHOT`
 
 Release date `Unreleased`
+
+- 
+
+## `1.0.57-beta7`
+
+Release date `2022-12-23`
 
 - Rename `handler/session-attachment` to `session`
 - Rename wrapper utility functions